reactoring adapters
