+++ main/adapters/encrypters/key-encrypter.go
@@ -1,4 +1,4 @@
-package handlers
+package encrypters
 
 import (
 	"fmt"
