+++ main/adapters/httphelper/resp.go
@@ -6,6 +6,12 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
+type HTTPResponse struct {
+	StatusCode int         `json:"statusCode"`
+	Header     http.Header `json:"header"`
+	Content    []byte      `json:"content"`
+}
+
 func EmptyOk(w http.ResponseWriter) {
 	w.Header().Set(HeaderContentType, MimeTextPlain)
 	w.WriteHeader(http.StatusOK)
@@ -22,4 +28,13 @@ func Ok(w http.ResponseWriter, rsp string) {
 	if err != nil {
 		log.Errorf("unable to write response: %s", err)
 	}
+}
+
+
+func HttpFailed(StatusCode int) bool {
+	return !HttpSuccess(StatusCode)
+}
+
+func HttpSuccess(StatusCode int) bool {
+	return StatusCode >= 200 && StatusCode < 300
 }
