+++ main/adapters/repository/init_db.go
@@ -1,4 +1,4 @@
-package handlers
+package repository
 
 import (
 	"context"
