--- main/repository/identity.go
@@ -1,6 +1,0 @@
-package repository
-
-type EntityRepo struct {
-	connStr string
-}
-
