+++ main/vars/vars.go
@@ -7,4 +7,17 @@ const (
 	PostgreSqlVersionTableName  string = "version"
 	MigrateArg                  string = "--migrate"
 	InitArg                     string = "--init-identities-conf"
+
+	UUIDKey      = "uuid"
+	OperationKey = "operation"
+	VerifyPath   = "verify"
+	HashEndpoint = "hash"
+
+	BinType  = "application/octet-stream"
+	TextType = "text/plain"
+	JSONType = "application/json"
+
+	HexEncoding = "hex"
+
+	HashLen = 32
 )
